+++ WinCopies.IO/FolderLoader.cs
@@ -42,7 +42,7 @@ namespace WinCopies.IO
         /// <summary>
         /// Initializes a new instance of the <see cref="BrowsableObjectInfoItemsLoader"/> class.
         /// </summary>
-        public FolderLoader(bool workerReportsProgress, bool workerSupportsCancellation, FileTypesFlags fileTypes) : base(workerReportsProgress, workerSupportsCancellation, fileTypes)
+        public FolderLoader(bool workerReportsProgress, bool workerSupportsCancellation, FileTypes fileTypes) : base(workerReportsProgress, workerSupportsCancellation, fileTypes)
 
         {
 
@@ -224,9 +224,9 @@ namespace WinCopies.IO
 
         {
 
-            if (FileTypes == FileTypesFlags.None) return;
+            if (FileTypes == FileTypes.None) return;
 
-            else if (FileTypes.HasFlag(FileTypesFlags.All) && FileTypes.HasMultipleFlags())
+            else if (FileTypes.HasFlag(FileTypes.All) && FileTypes.HasMultipleFlags())
 
                 throw new InvalidOperationException("FileTypes cannot have the All flag in combination with other flags.");
 
@@ -262,7 +262,7 @@ namespace WinCopies.IO
 
                     Debug.WriteLine("");
 
-                if (pathInfo.FileType == FileType.None || (pathInfo.FileType != FileType.SpecialFolder && FileTypes != FileTypesFlags.All && !FileTypes.HasFlag(FileTypeToFileTypeFlags(pathInfo.FileType)))) return;
+                if (pathInfo.FileType == FileType.None || (pathInfo.FileType != FileType.SpecialFolder && FileTypes != FileTypes.All && !FileTypes.HasFlag(FileTypeToFileTypeFlags(pathInfo.FileType)))) return;
 
                 // We only make a normalized path if we add the path to the paths to load.