+++ WinCopies.Util/BackgroundWorker.cs
@@ -7,33 +7,33 @@ using static WinCopies.Util.Generic;
 namespace WinCopies.Util
 {
 
+    ///// <para>FR: Représente un BackgroundWorker qui s'exécute par défaut dans un thread MTA et qui arrête automatiquement l'exécution en arrière-plan lors d'un rapport du progrès.</para>
     /// <summary>
-    /// <para>FR: Représente un BackgroundWorker qui s'exécute par défaut dans un thread MTA et qui arrête automatiquement l'exécution en arrière-plan lors d'un rapport du progrès.</para>
-    /// <para>EN: Represents a BackgroundWorker that runs in a MTA thread by default and automatically stops on background when reports progress.</para>
+    /// Represents a BackgroundWorker that runs in a MTA thread by default and automatically stops on background when reports progress.
     /// </summary>
-    public class BackgroundWorker : Component
+    public class BackgroundWorker : Component, IBackgroundWorker
     {
 
+        ///// <para>FR: Cet evènement se produit lorsque le thread d'arrière plan démarre. Placez votre code de traitement ici.</para>
+        ///// <para>Gestionnaire d'événement exécuté dans le thread d'arrière plan.</para>
         /// <summary>
-        /// <para>FR: Cet evènement se produit lorsque le thread d'arrière plan démarre. Placez votre code de traitement ici.</para>
-        /// <para>Gestionnaire d'événement exécuté dans le thread d'arrière plan.</para>
-        /// <para>EN: This event is called when the background thread starts. Put your background working code here.</para>
+        /// <para>This event is called when the background thread starts. Put your background working code here.</para>
         /// <para>The event handler is running in the background thread.</para>
         /// </summary>
         public event DoWorkEventHandler DoWork;
 
+        ///// <para>FR: Cet évènement se produit lorsque le thread d'arrière plan notifie de la progression.</para>
+        ///// <para>Gestionnaire d'événement exécuté dans le thread principal.</para>
         /// <summary>
-        /// <para>FR: Cet évènement se produit lorsque le thread d'arrière plan notifie de la progression.</para>
-        /// <para>Gestionnaire d'événement exécuté dans le thread principal.</para>
-        /// <para>EN: This event is called when the background thread reports progress.</para>
+        /// <para>This event is called when the background thread reports progress.</para>
         /// <para>The event handler is running in the main thread.</para>
         /// </summary>
         public event ProgressChangedEventHandler ProgressChanged;
 
+        ///// <para>FR: Cet évènement se produit lorsque le thread d'arrière plan est terminé.</para>
+        ///// <para>Gestionnaire d'événement exécuté dans le thread d'arrière plan.</para>
         /// <summary>
-        /// <para>FR: Cet évènement se produit lorsque le thread d'arrière plan est terminé.</para>
-        /// <para>Gestionnaire d'événement exécuté dans le thread d'arrière plan.</para>
-        /// <para>EN: This event is called when the background thread has finished working.</para>
+        /// <para>This event is called when the background thread has finished working.</para>
         /// <para>The event handler is running in the background thread.</para>
         /// </summary>
         public event RunWorkerCompletedEventHandler RunWorkerCompleted;
@@ -44,62 +44,50 @@ namespace WinCopies.Util
 
         private ApartmentState _ApartmentState = ApartmentState.MTA;
 
-        private SynchronizationContext _SyncContext;
+        private readonly SynchronizationContext _SyncContext;
 
-        private ManualResetEvent _event = new ManualResetEvent(true);
+        private readonly ManualResetEvent _event = new ManualResetEvent(true);
 
 
 
+        ///// <para>FR: Obtient une valeur indiquant si le traitement a été annulé.</para>
         /// <summary>
-        /// <para>FR: Obtient une valeur indiquant si le traitement a été annulé.</para>
-        /// <para>EN: Gets a value that indicates if the working is cancelled.</para>
+        /// Gets a value that indicates whether the working has been cancelled.
         /// </summary>
         public bool IsCancelled { get; private set; } = false;
 
+        ///// <para>FR: Obtient une valeur indiquant si le thread doit essayer de se terminer avant la fin de ses tâches en arrière-plan.</para>
         /// <summary>
-        /// <para>FR: Obtient une valeur indiquant si le thread doit essayer de se terminer avant la fin de ses tâches en arrière-plan.</para>
-        /// <para>EN: Gets a value that indicates if the thread must try to cancel before finished the background tasks.</para>
+        /// Gets a value that indicates whether the thread must try to cancel before finished the background tasks.
         /// </summary>
         public bool CancellationPending { get; private set; } = false;
 
+        ///// <para>FR: Obtient une valeur indiquant si le thread est occupé.</para>
         /// <summary>
-        /// <para>FR: Obtient une valeur indiquant si le thread est occupé.</para>
-        /// <para>EN: Gets a value that indicates if the thread is busy.</para>
+        /// Gets a value that indicates whether the thread is busy.
         /// </summary>
         public bool IsBusy { get; private set; } = false;
 
-        private bool workerReportsProgress = false;
+        private readonly bool workerReportsProgress = false;
 
+        ///// <para>FR: Obtient ou définit une valeur indiquant si le thread peut notifier de l'avancement.</para>
         /// <summary>
-        /// <para>FR: Obtient ou définit une valeur indiquant si le thread peut notifier de l'avancement.</para>
-        /// <para>EN: Gets or sets a value that indicates if the thread can notify of the progress.</para>
+        /// Gets or sets a value that indicates whether the thread can notify of the progress.
         /// </summary>
         public bool WorkerReportsProgress
         {
 
             get => workerReportsProgress;
 
-            set
-
-            {
-
-                if (IsBusy)
-
-                    throw new InvalidOperationException(BackgroundWorkerIsBusy);
-
-                else
-
-                    workerReportsProgress = value;
-
-            }
+            set => this.SetBackgroundWorkerProperty(nameof(WorkerReportsProgress), nameof(workerReportsProgress), value, typeof(BackgroundWorker), true);
 
         }
 
-        private bool workerSupportsCancellation = false;
+        private readonly bool workerSupportsCancellation = false;
 
+        ///// <para>FR: Obtient ou définit une valeur indiquant si le thread supporte l'annulation.</para>
         /// <summary>
-        /// <para>FR: Obtient ou définit une valeur indiquant si le thread supporte l'annulation.</para>
-        /// <para>EN: Gets or sets a value that indicates if the thread supports the cancellation.</para>
+        /// Gets or sets a value that indicates whether the thread supports the cancellation.
         /// </summary>
         public bool WorkerSupportsCancellation
 
@@ -107,31 +95,19 @@ namespace WinCopies.Util
 
             get => workerSupportsCancellation;
 
-            set
-
-            {
-
-                if (IsBusy)
-
-                    throw new InvalidOperationException(BackgroundWorkerIsBusy);
-
-                else
-
-                    workerSupportsCancellation = true;
-
-            }
+            set => this.SetBackgroundWorkerProperty(nameof(WorkerReportsProgress), nameof(workerSupportsCancellation), value, typeof(BackgroundWorker), true);
 
         }
 
+        ///// <para>FR: Obtient le progrès actuel du traitement en pourcents.</para>
         /// <summary>
-        /// <para>FR: Obtient le progrès actuel du traitement en pourcents.</para>
-        /// <para>EN: Gets the current progress of the working in percent.</para>
+        /// Gets the current progress of the working in percent.
         /// </summary>
         public int Progress { get; private set; } = 0;
 
+        ///// <para>FR: Obtient l'<see cref="System.Threading. ApartmentState"/> de ce thread.</para>
         /// <summary>
-        /// <para>FR: Obtient l'<see cref="System.Threading. ApartmentState"/> de ce thread.</para>
-        /// <para>EN: Gets the <see cref="System.Threading.ApartmentState"/> of this thread.</para>
+        /// Gets the <see cref="System.Threading.ApartmentState"/> of this thread.
         /// </summary>
         public ApartmentState ApartmentState
         {
@@ -141,25 +117,27 @@ namespace WinCopies.Util
             set
             {
 
-                if (IsBusy)
-
-                    // todo:
-
-                    throw new InvalidOperationException("Le thread est déjà en cours d'exécution ; son ApartmentState ne peut pas encore changer.");
-
-                else if (_ApartmentState == value)
+                if (_ApartmentState == value)
 
                     return;
 
-                // todo:
-
                 else if (value == ApartmentState.Unknown)
 
-                    throw new ArgumentException(string.Format("La valeur {0} ne peut pas être sélectionnée pour un ApartmentState d'un {1}.", ApartmentState.Unknown.GetType().FullName, typeof(BackgroundWorker).FullName));
+                    // todo:
 
-                else
+                    //La valeur {0} ne peut pas être sélectionnée pour un ApartmentState d'un {1}.
 
-                    _ApartmentState = value;
+                    throw new ArgumentException(string.Format("The {0} value is not valid for the {1} class.", nameof(ApartmentState.Unknown), nameof(BackgroundWorker)));
+
+                else if (IsBusy)
+
+                    // todo:
+
+                    //Le thread est déjà en cours d'exécution ; son ApartmentState ne peut pas encore changer.
+
+                    throw new InvalidOperationException(BackgroundWorkerIsBusy);
+
+                _ApartmentState = value;
 
             }
 
@@ -167,9 +145,9 @@ namespace WinCopies.Util
 
 
 
+        ///// <para>FR: Construit un <see cref="BackgroundWorker"/>.</para>
         /// <summary>
-        /// <para>FR: Construit un <see cref="BackgroundWorker"/>.</para>
-        /// <para>EN: Initializes a new instance of the <see cref="BackgroundWorker"/> class.</para>
+        /// Initializes a new instance of the <see cref="BackgroundWorker"/> class.
         /// </summary>
         public BackgroundWorker() => _SyncContext = SynchronizationContext.Current ?? new SynchronizationContext();
         // {
@@ -182,13 +160,13 @@ namespace WinCopies.Util
 
         // }
 
+        ///// <para>FR: Construit un <see cref="BackgroundWorker"/> avec un <see cref="System.Threading.ApartmentState"/> donné.</para>
+        ///// <para>FR: <see cref="System.Threading.ApartmentState"/> dans lequel initialiser le thread.</para>
         /// <summary>
-        /// <para>FR: Construit un <see cref="BackgroundWorker"/> avec un <see cref="System.Threading.ApartmentState"/> donné.</para>
-        /// <para>EN: Initializes a new instance of the <see cref="BackgroundWorker"/> class with a given <see cref="System.Threading.ApartmentState"/>.</para>
+        /// Initializes a new instance of the <see cref="BackgroundWorker"/> class with a given <see cref="System.Threading.ApartmentState"/>.
         /// </summary>
         /// <param name="apartmentState">
-        /// <para>FR: <see cref="System.Threading.ApartmentState"/> dans lequel initialiser le thread.</para>
-        /// <para>EN: <see cref="System.Threading.ApartmentState"/> in which initialize the thread.</para>
+        /// The <see cref="System.Threading.ApartmentState"/> in which to initialize the thread.
         /// </param>
         public BackgroundWorker(ApartmentState apartmentState)
         {
@@ -207,9 +185,9 @@ namespace WinCopies.Util
 
 
 
+        ///// <para>FR: Ré-initialise les variables locales.</para>
         /// <summary>
-        /// <para>FR: Ré-initialise les variables locales.</para>
-        /// <para>EN: Re-initializes the local variables.</para>
+        /// Re-initializes the local variables.
         /// </summary>
         private void Reset(bool isCancelled)
         {
@@ -228,19 +206,19 @@ namespace WinCopies.Util
 
         // DoWorkEventArgs e = null;
 
+        ///// <para>FR: Démarre le traitement.</para>
         /// <summary>
-        /// <para>FR: Démarre le traitement.</para>
-        /// <para>EN: Starts the working.</para>
+        /// Starts the working.
         /// </summary>
         public void RunWorkerAsync() => RunWorkerAsync(null);
 
+        ///// <para>FR: Démarre le traitement avec un argument personnalisé.</para>
+        ///// <para>FR: Argument passé au traitement.</para>
         /// <summary>
-        /// <para>FR: Démarre le traitement avec un argument personnalisé.</para>
-        /// <para>EN: Starts the working with a custom parameter.</para>
+        /// Starts the working with a custom parameter.
         /// </summary>
         /// <param name="argument">
-        /// <para>FR: Argument passé au traitement.</para>
-        /// <para>EN: Argument given for the working.</para>
+        /// Argument given for the working.
         /// </param>
         public void RunWorkerAsync(object argument)
         {
@@ -271,9 +249,9 @@ namespace WinCopies.Util
 
         }
 
+        ///// <para>FR: Point d'entré du thread.</para>
         /// <summary>
-        /// <para>FR: Point d'entré du thread.</para>
-        /// <para>EN: Entry point of the thread.</para>
+        /// Entry point of the thread.
         /// </summary>
         // /// <param name="argument">Argument du thread.</param>
         private void ThreadStart(DoWorkEventArgs e)
@@ -305,13 +283,17 @@ namespace WinCopies.Util
 
             }
 
-            _SyncContext.Send(ThreadCompleted, new ValueTuple<object, Exception, bool>(e.Result, error, IsCancelled || CancellationPending || e.Cancel));
+            bool isCancelled = IsCancelled || CancellationPending || e.Cancel;
+
+            Reset(isCancelled);
+
+            _SyncContext.Send(ThreadCompleted, new ValueTuple<object, Exception, bool>(e.Result, error, isCancelled ));
 
         }
 
+        ///// <para>FR: Cette méthode est appelée lorsque le thread est terminé.</para>
         /// <summary>
-        /// <para>FR: Cette méthode est appelée lorsque le thread est terminé.</para>
-        /// <para>EN: This method is called when the thread has finished.</para>
+        /// The method that is called when the thread has finished.
         /// </summary>
         // /// <param name="result">Objet résultat du traitement.</param>
         // /// <param name="error">Erreur éventuelle. <c>null</c> si pas d'erreur.</param>
@@ -321,7 +303,7 @@ namespace WinCopies.Util
 
             (object result, Exception ex, bool isCancelled) = (ValueTuple<object, Exception, bool>)args;
 
-            RunWorkerCompletedEventArgs e = new RunWorkerCompletedEventArgs(result, ex, isCancelled);
+            var e = new RunWorkerCompletedEventArgs(result, ex, isCancelled);
 
             // IsBusy = false;
 
@@ -329,25 +311,21 @@ namespace WinCopies.Util
 
             // IsCancelled = r.Item3;    
 
-            Reset(isCancelled);
-
             RunWorkerCompleted?.Invoke(this, e);
 
         }
 
+        ///// <para>FR: Annule le traitement en cours de manière asynchrone.</para>
         /// <summary>
-        /// <para>FR: Annule le traitement en cours de manière asynchrone.</para>
-        /// <para>EN: Cancels the working asynchronously.</para>
+        /// Cancels the working asynchronously.
         /// </summary>
         public void CancelAsync() => Cancel(false);
 
+        ///// <para>FR: Annule le traitement en cours.</para>
         /// <summary>
-        /// <para>FR: Annule le traitement en cours.</para>
-        /// <para>EN: Cancels the working.</para>
+        /// Cancels the working.
         /// </summary>
         public void Cancel() => Cancel(true);
-
-
 
         private void Cancel(bool abort)
         {
@@ -356,7 +334,9 @@ namespace WinCopies.Util
 
                 // todo:
 
-                throw new InvalidOperationException("Ce traitement ne supporte pas l'annulation.");
+                //Ce traitement ne supporte pas l'annulation.
+
+                throw new InvalidOperationException("This BackgroundWorker does not support cancellation.");
 
 
 
@@ -381,76 +361,101 @@ namespace WinCopies.Util
 
         }
 
+        ///// <para>FR: Délégué pour rapporter de la progression.</para>
+        ///// <para>FR: Argument de l'évènement.</para>
         /// <summary>
-        /// <para>FR: Délégué pour rapporter de la progression.</para>
-        /// <para>EN: Delegate for progress reportting.</para>
+        /// Delegate for progress reportting.
         /// </summary>
         /// <param name="args">
-        /// <para>FR: Argument de l'évènement.</para>
-        /// <para>EN: Event argument.</para>
+        /// Event argument.
         /// </param>
         private void OnProgressChanged(object args) => ProgressChanged?.Invoke(this, args as ProgressChangedEventArgs);
 
+        ///// <para>FR: Notifie de la progession.</para>
+        ///// <para>FR: Pourcentage de progression.</para>
         /// <summary>
-        /// <para>FR: Notifie de la progession.</para>
-        /// <para>EN: Notifies of the progress.</para>
+        /// Notifies of the progress.
         /// </summary>
         /// <param name="percentProgress">
-        /// <para>FR: Pourcentage de progression.</para>
-        /// <para>EN: Progress percentage.</para>
+        /// Progress percentage.
         /// </param>
         public void ReportProgress(int percentProgress) => ReportProgress(percentProgress, null);
 
+        ///// <para>FR: Notifie de la pogression.</para>
+        ///// <para>FR: Pourcentage de progression.</para>
+        ///// <para>FR: Objet utilisateur.</para>
         /// <summary>
-        /// <para>FR: Notifie de la pogression.</para>
-        /// <para>EN: Notifies of the progress.</para>
+        /// Notifies of the progress.
         /// </summary>
         /// <param name="percentProgress">
-        /// <para>FR: Pourcentage de progression.</para>
-        /// <para>EN: Progress percentage.</para>
+        /// Progress percentage.
         /// </param>
         /// <param name="userState">
-        /// <para>FR: Objet utilisateur.</para>
-        /// <para>EN: User object.</para>
+        /// User object.
         /// </param>
         public void ReportProgress(int percentProgress, object userState)
         {
 
             if (!WorkerReportsProgress)
 
-                throw new InvalidOperationException("Ce BackgroundWorker ne permet pas de notifier de la progression.");
+                //Ce BackgroundWorker ne permet pas de notifier de la progression.
+
+                throw new InvalidOperationException("This BackgroundWorker does not support progression notification.");
 
             Progress = percentProgress;
 
-            ProgressChangedEventArgs e = new ProgressChangedEventArgs(Progress, userState);
+            var e = new ProgressChangedEventArgs(Progress, userState);
 
             _SyncContext.Send(OnProgressChanged, e);
 
         }
 
+        /// <summary>
+        /// Suspends the current thread.
+        /// </summary>
         public void Suspend()
 
         {
 
-            // to suspend thread.
-            _event.Reset();
+            _ = _event.Reset();
 
-            _event.WaitOne();
+            _ = _event.WaitOne();
 
         }
 
+        /// <summary>
+        /// Resumes the current thread.
+        /// </summary>
         public void Resume() =>
 
-            //to resume thread
             _event.Set();
 
+        /// <summary>
+        /// Gets a value that indicates whether the current <see cref="BackgroundWorker"/> is disposed.
+        /// </summary>
+        public bool IsDisposed { get; private set; }
+
+        /// <summary>
+        /// Releases the unmanaged resources used by the <see cref="BackgroundWorker"/> and optionally releases the managed resources.
+        /// </summary>
+        /// <param name="disposing"><see langword="true"/> to release both managed and unmanaged resources; <see langword="false"/> to release only unmanaged resources.</param>
         protected override void Dispose(bool disposing)
 
         {
 
+            if (IsBusy)
+
+                throw new InvalidOperationException(BackgroundWorkerIsBusy);
+
             base.Dispose(disposing);
 
+            if (IsDisposed)
+
+                return;
+
             _event.Dispose();
+
+            IsDisposed = true;
 
         }